Refactor qedrv.c: extract ch_io to an asm frontend stub to gain full control over registers.
Couldn't find a way to have proper control on which registers are saved
when QDOS calls ch_io. This is why ch_io now is an asm stub that calls
the C implementations.

Also started refactoring towards getting rid of socket.c and use direct W5300
access in chan_ops.c. This makes the code easier to reason about and reduces
need for buffer copying.